Refactor: Extract terminal launcher
